# swagger.json
swagger: '2.0'
info:
  version: "1.0.1"
  title: Wardrobe
  description: Simple UI to promote .war files between S3 locations
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
basePath: "/v1"
...
schemes:
- "http"
paths:
  /shelf:
    post: 
      tags:
      - "shelf"
      produces:
      - "application/xml"
      - "application/json"
      consumes:
      - "application/xml"
      - "application/json"
      operationId: "addShelf"
      parameters:
      - in: "body"
        name: "body"
        description: "Shelf object that needs to be added to the wardrobe"
        required: true
        schema:
          $ref: "#/definitions/Shelf"
      responses:
        405:
          description: "Invalid input"
  /shelf/{shelfId}:
    get: 
      tags:
      - "shelf"
      summary: "Find shelf by ID"
      description: "Returns a single shelf"
      operationId: "getShelfById"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "shelfId"
        in: "path"
        description: "ID of shelf to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Shelf"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Shelf not found"
    put:
      tags:
      - "shelf"
      summary: "Update an existing shelf"
      description: ""
      operationId: "updateShelf"
      produces:
      - "application/xml"
      - "application/json"
      consumes:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "shelfId"
        in: "path"
        description: "ID of shelf to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Shelf not found"
        405:
          description: "Validation exception"
    delete:
      tags:
      - "shelf"
      summary: "Update an existing shelf"
      description: ""
      operationId: "deleteShelf"
      produces:
      - "application/xml"
      - "application/json"
      consumes:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "shelfId"
        in: "path"
        description: "ID of shelf to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        405:
          description: "Invalid input"
  /shelf/{shelfId}/promote/{promoterId}:
    get:
      tags:
      - "shelf"
      summary: "Find shelf by ID"
      description: "Adds a new revision defined by the promoter"
      operationId: "promoteDrawerId"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "shelfId"
        in: "path"
        description: "ID of shelf to promote"
        required: true
        type: "integer"
        format: "int64"
      - name: "promoterId"
        in: "path"
        description: "ID of promoter to promote from"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
  /shelf/{shelfId}/rollback/{revisionId}:
    get:
      tags:
      - "shelf"
      summary: "Find shelf by ID"
      description: "Sets the file to another version"
      operationId: "rollbackToDrawerId"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "shelfId"
        in: "path"
        description: "ID of shelf to rollback"
        required: true
        type: "integer"
        format: "int64"
      - name: "revisionId"
        in: "path"
        description: "ID of revision to rollback to"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
...
  /stack/:
      get:
        tags:
        - "stack"
        summary: "Get stacks"
        description: "."
        operationId: "getStacks"
        produces:
        - "application/xml"
        - "application/json"
        responses:
          200:
            description: "successful operation"
            schema:
              type: "array"
              items:
                $ref: "#/definitions/Stack"
      post: 
        tags:
        - "stack"
        summary: "Create stack"
        description: "."
        operationId: "createStack"
        consumes:
        - "application/xml"
        - "application/json"
        produces:
        - "application/xml"
        - "application/json"
        parameters:
        - in: "body"
          name: "body"
          description: "Stack object that needs to be added to a shelf"
          required: true
          schema:
            $ref: "#/definitions/Stack"
        responses:
          405:
            description: "Invalid input"
  /stack/{stackId}:
    get: 
      tags:
      - "stack"
      parameters:
      - name: "stackId"
        in: "path"
        description: "ID of stack to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Stack"
        400:
          description: "Invalid ID supplied"
        404:
          description: "stack not found"
    put:
      tags:
      - "stack"
      parameters:
      - name: "stackId"
        in: "path"
        description: "ID of shelf to update"
        required: true
        type: "integer"
        format: "int64"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "stack not found"
        405:
          description: "Validation exception"
    delete:
      tags:
      - "stack"
      parameters:
      - name: "stackId"
        in: "path"
        description: "ID of stack to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        405:
          description: "Invalid input"
...
definitions:
  Stack:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      name:
        type: "string"
      authGroup:
        type: "string"
      shelves:
        type: "array"
        items:
          $ref: "#/definitions/Shelf"
    xml:
      name: "Stack"
  Revision:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int32"
      dateAdded:
        type: "string"
        format: "date"
    xml:
      name: "Revision"
  Shelf:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      url:
        type: "string"
      promoters:
        type: "array"
        items:
          $ref: "#/definitions/Promoter"
      revisions:
        type: "array"
        items:
          $ref: "#/definitions/Revision"
      latest:
        $ref: "#/definitions/Revision"
    xml:
      name: "Shelf"
  Promoter:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      name:
        type: "string"
      address:
        type: "string"
    xml:
      name: "Promoter"
...